+++ src/components/form/CreateModelForm.vue
@@ -4,22 +4,6 @@
     data-cy="create-model-form"
     @submit="onSubmit"
   >
-    <q-input
-      v-model="modelName"
-      filled
-      :label="$t('actions.models.create.form.name')"
-      lazy-rules
-      :rules="[
-        (value) => notEmpty($t, value),
-        (value) => isUnique(
-          $t,
-          models.map((model) => `${model.plugin}/${model.name}`),
-          `${modelPlugin}/${value}`,
-          'errors.models.duplicate',
-        )
-      ]"
-      data-cy="name-input"
-    />
     <q-select
       v-model="modelPlugin"
       filled
@@ -29,6 +13,7 @@
         (value) => notEmpty($t, value),
       ]"
       data-cy="plugin-select"
+      @update:model-value="onPluginChange"
     >
       <template #option="{ selected, opt, toggleOption }">
         <q-item
@@ -42,6 +27,31 @@
         </q-item>
       </template>
     </q-select>
+    <q-input
+      v-model="modelName"
+      filled
+      :label="$t('actions.models.create.form.name')"
+      lazy-rules
+      :rules="[
+        (value) => notEmpty($t, value),
+        (value) => isUniqueModel(
+          $t,
+          modelPlugin,
+          models.filter(({ plugin }) => plugin === modelPlugin).map(({ path }) => path),
+          `${baseFolder}${modelName}`,
+          'errors.models.duplicate',
+        )
+      ]"
+      data-cy="name-input"
+    />
+    <q-input
+      v-model="modelName"
+      :model-value="`${baseFolder}${modelName}`"
+      outlined
+      disable
+      :label="$t('actions.models.create.form.location')"
+      data-cy="location-input"
+    />
     <div class="flex row items-center justify-center">
       <q-btn
         icon="fa-solid fa-save"
@@ -61,12 +71,22 @@
 
 <script setup>
 import { Notify } from 'quasar';
-import { getPlugins } from 'src/composables/PluginManager';
-import { onMounted, reactive, ref } from 'vue';
-import { isUnique, notEmpty } from 'src/composables/QuasarFieldRule';
+import { getPluginByName, getPlugins } from 'src/composables/PluginManager';
+import {
+  computed,
+  onMounted,
+  reactive,
+  ref,
+} from 'vue';
+import { isUniqueModel, notEmpty } from 'src/composables/QuasarFieldRule';
 import { useI18n } from 'vue-i18n';
-import { createProjectFolder, getAllModels } from 'src/composables/Project';
+import {
+  appendProjectFile,
+  getAllModels,
+  getProjectFolders,
+} from 'src/composables/Project';
 import { useRouter } from 'vue-router';
+import { FileInformation, FileInput } from 'leto-modelizer-plugin-core';
 
 const { t } = useI18n();
 const router = useRouter();
@@ -79,10 +99,12 @@ const props = defineProps({
 });
 
 const plugins = reactive(getPlugins());
-const modelName = ref(null);
+const modelName = ref(plugins[0].configuration.defaultFileName);
 const modelPlugin = ref(plugins[0]?.data.name);
 const submitting = ref(false);
 const models = ref([]);
+const allFolders = ref([]);
+const baseFolder = computed(() => getPluginByName(modelPlugin.value).configuration.restrictiveFolder || '');
 
 /**
  * Create a new model folder and its parent folders if necessary.
@@ -91,11 +113,15 @@ const models = ref([]);
  * @return {Promise<void>} Promise with nothing on success or error.
  */
 async function onSubmit() {
-  const pluginFolder = process.env.MODELS_DEFAULT_FOLDER !== ''
-    ? `${process.env.MODELS_DEFAULT_FOLDER}/${modelPlugin.value}`
-    : `${modelPlugin.value}`;
+  const diagram = `${baseFolder.value}${modelName.value || ''}`;
+  const model = getPluginByName(modelPlugin.value)
+    .getModels([new FileInformation({ path: diagram })])
+    .find(() => true);
 
-  return createProjectFolder(props.projectName, `${pluginFolder}/${modelName.value}`)
+  return appendProjectFile(
+    props.projectName,
+    new FileInput({ path: diagram, content: '' }),
+  )
     .then(() => {
       Notify.create({
         type: 'positive',
@@ -108,7 +134,10 @@ async function onSubmit() {
         params: {
           projectName: props.projectName,
         },
-        query: { path: `${modelPlugin.value}/${modelName.value}` },
+        query: {
+          plugin: modelPlugin.value,
+          path: model,
+        },
       });
     })
     .catch((error) => {
@@ -130,9 +159,17 @@ async function onSubmit() {
     });
 }
 
-onMounted(() => {
+/**
+ * Set default model name on plugin name change.
+ */
+function onPluginChange() {
+  modelName.value = getPluginByName(modelPlugin.value).configuration.defaultFileName;
+}
+
+onMounted(async () => {
   getAllModels(props.projectName).then((array) => {
     models.value = array;
   });
+  allFolders.value = await getProjectFolders(props.projectName);
 });
 </script>