+++ src/composables/PluginManager.js
@@ -1,6 +1,6 @@
 import { readTextFile } from 'src/composables/Files';
 import plugins from 'src/plugins';
-import { FileInput, FileInformation } from 'leto-modelizer-plugin-core';
+import { FileInformation, FileInput } from 'leto-modelizer-plugin-core';
 import {
   writeProjectFile,
   readProjectFile,
@@ -8,10 +8,12 @@ import {
   deleteProjectFile,
   readDir,
   appendProjectFile,
+  isDirectory,
 } from 'src/composables/Project';
 import PluginEvent from 'src/composables/events/PluginEvent';
 import { getTemplateFiles } from 'src/composables/TemplateManager';
 
+const configurationFilePath = './leto-modelizer.config.json';
 let instanciatePlugins = [];
 
 /**
@@ -133,7 +135,12 @@ export function getPluginTags(name) {
  * @return {Object} Return an array of tags.
  */
 export function getAllTags() {
-  return [...new Set(getPlugins().flatMap(({ configuration }) => configuration.tags))].sort();
+  return [...new Set(
+    getPlugins()
+      .flatMap(({ configuration }) => configuration.tags)
+      .filter(({ type }) => type === 'category')
+      .map(({ value }) => value),
+  )].sort();
 }
 
 /**
@@ -150,19 +157,26 @@ export function isParsableFile(file) {
  * Render the given model with the corresponding pugin.
  * Return rendered files.
  * @param {String} projectId - ID of the project.
- * @param {String} modelPath - Path of the models folder.
+ * @param {String} modelPath - Path of the model folder.
  * @param {Object} plugin - Plugin to render.
  * @return {Promise<Array<FileInput>>} Promise with FileInputs array on success otherwise an error.
  */
 export async function renderModel(projectId, modelPath, plugin) {
-  const config = new FileInput({
-    path: `${modelPath}/leto-modelizer.config.json`,
-    content: '{}',
-  });
+  const isFolder = await isDirectory(`${projectId}/${modelPath}`);
+  const modelFolder = isFolder ? modelPath : modelPath.substring(0, modelPath.lastIndexOf('/'));
 
-  const files = await getModelFiles(projectId, modelPath, plugin);
+  const config = await readProjectFile(
+    projectId,
+    new FileInformation({
+      path: configurationFilePath,
+    }),
+  );
+  const files = isFolder
+    ? await getModelFiles(projectId, modelFolder, plugin)
+    : [new FileInput({ path: modelPath })];
 
   const renderFiles = plugin.render(
+    new FileInformation({ path: modelFolder }),
     config,
     files.filter((file) => plugin.isParsable(file)),
   );
@@ -181,19 +195,21 @@ export async function renderModel(projectId, modelPath, plugin) {
 /**
  * Render the configuration file.
  * @param {String} projectId - ID of the project.
- * @param {String} modelPath - Path of the models folder.
+ * @param {String} modelPath - Path of the model.
  * @param {Object} plugin - Plugin to render.
  * @return {Promise<void>} Promise with nothing on success otherwise an error.
  */
 export async function renderConfiguration(projectId, modelPath, plugin) {
-  const config = new FileInput({
-    path: `${modelPath}/leto-modelizer.config.json`,
-    content: '{}',
-  });
+  const config = await readProjectFile(
+    projectId,
+    new FileInformation({
+      path: configurationFilePath,
+    }),
+  );
 
   // TODO : replace by appropriate function when it's done in plugin-core
   // eslint-disable-next-line no-underscore-dangle
-  plugin.__renderer.renderConfiguration(config);
+  plugin.__renderer.renderConfiguration(new FileInformation({ path: modelPath }), config);
 
   await writeProjectFile(projectId, config);
 }
@@ -224,24 +240,30 @@ export async function getFileInputs(plugin, fileInformations, projectName) {
  * @return {Promise<void>} Promise with nothing on success otherwise an error.
  */
 export async function initComponents(projectName, plugin, path) {
-  const dir = path;
-  const files = await readDir(`${projectName}/${dir}`);
-  const fileInformations = files.map((file) => new FileInformation({ path: `${dir}/${file}` }));
+  let filesInformation;
 
-  const fileInputs = await getFileInputs(plugin, fileInformations, projectName);
+  if (await isDirectory(`${projectName}/${path}`)) {
+    filesInformation = (await readDir(`${projectName}/${path}`))
+      .map((file) => new FileInformation({ path: `${path}/${file}` }))
+      .filter((file) => plugin.isParsable(file));
+  } else {
+    filesInformation = [new FileInformation({ path })];
+  }
+
+  const fileInputs = await getFileInputs(plugin, filesInformation, projectName);
 
   const config = await readProjectFile(
     projectName,
     new FileInformation({
-      path: `${dir}/leto-modelizer.config.json`,
+      path: configurationFilePath,
     }),
   );
 
-  plugin.parse(config, fileInputs);
+  plugin.parse(new FileInformation({ path }), config, fileInputs);
 }
 
 /**
- * Add a new component and render model.
+ * Add a new component.
  * @param {String} projectName - Name of the project.
  * @param {Object} plugin - Plugin corresponding to the model.
  * @param {String} path - Model path (Plugin name & model name).
@@ -256,14 +278,17 @@ export async function addNewComponent(
   definition,
   drawOption = null,
 ) {
-  const componentId = plugin.data.addComponent(definition, `${path}/`);
+  let componentId;
 
-  if (drawOption) {
-    const component = plugin.data.getComponentById(componentId);
-    component.drawOption = drawOption;
+  if (await isDirectory(`${projectName}/${path}`)) {
+    componentId = plugin.data.addComponent(definition, `${path}/`);
+  } else {
+    componentId = plugin.data.addComponent(definition, '', path);
   }
 
-  await renderModel(projectName, path, plugin);
+  const component = plugin.data.getComponentById(componentId);
+
+  component.drawOption = drawOption;
 }
 
 /**
@@ -301,9 +326,19 @@ export async function addNewTemplateComponent(
   const config = await readProjectFile(
     projectName,
     new FileInformation({
-      path: `${path}/leto-modelizer.config.json`,
+      path: configurationFilePath,
     }),
   );
 
-  plugin.parse(config, fileInputs);
+  plugin.parse(new FileInformation({ path }), config, fileInputs);
+}
+
+/**
+ * Get path of model corresponding to the file path.
+ * @param {Object} plugin - Plugin corresponding to the model.
+ * @param {String} path - File path.
+ * @return {String} Model path.
+ */
+export function getModelPath(plugin, path) {
+  return getPluginByName(plugin).getModels([new FileInformation({ path })])[0];
 }