+++ tests/unit/components/FileExplorer.spec.js
@@ -14,6 +14,7 @@ installQuasarPlugin();
 jest.mock('src/composables/Project', () => ({
   getProjectFiles: jest.fn(() => Promise.resolve([])),
   getStatus: jest.fn(() => Promise.resolve([])),
+  isDirectory: jest.fn((file) => Promise.resolve(file === 'folder')),
 }));
 
 jest.mock('vue-router', () => ({
@@ -515,21 +516,62 @@ describe('Test component: FileExplorer', () => {
   });
 
   describe('Test function: openModelFiles', () => {
-    it('should call getPluginByName()', () => {
+    it('should do nothing without plugin', async () => {
       const mock = jest.fn(() => ({
         isParsable: () => true,
       }));
-      PluginManager.getPluginByName.mockImplementation(mock);
 
-      process.env.MODELS_DEFAULT_FOLDER = 'default';
+      PluginManager.getPluginByName.mockImplementation(mock);
 
       wrapper.vm.fileExplorerRef = {
         getNodeByKey: jest.fn(() => ({ children: [] })),
         setExpanded: jest.fn(),
       };
       wrapper.vm.query.path = 'pluginName/modelName';
+      wrapper.vm.query.plugin = '';
       wrapper.vm.localFileInformations = [{ path: 'pluginName/modelName/file.ext' }];
-      wrapper.vm.openModelFiles();
+
+      await wrapper.vm.openModelFiles();
+
+      expect(mock).toHaveBeenCalledTimes(0);
+    });
+
+    it('should do nothing without model', async () => {
+      const mock = jest.fn(() => ({
+        isParsable: () => true,
+      }));
+
+      PluginManager.getPluginByName.mockImplementation(mock);
+
+      wrapper.vm.fileExplorerRef = {
+        getNodeByKey: jest.fn(() => ({ children: [] })),
+        setExpanded: jest.fn(),
+      };
+      wrapper.vm.query.path = '';
+      wrapper.vm.query.plugin = 'test';
+      wrapper.vm.localFileInformations = [{ path: 'pluginName/modelName/file.ext' }];
+
+      await wrapper.vm.openModelFiles();
+
+      expect(mock).toHaveBeenCalledTimes(0);
+    });
+
+    it('should call getPluginByName()', async () => {
+      const mock = jest.fn(() => ({
+        isParsable: () => true,
+      }));
+
+      PluginManager.getPluginByName.mockImplementation(mock);
+
+      wrapper.vm.fileExplorerRef = {
+        getNodeByKey: jest.fn(() => ({ children: [] })),
+        setExpanded: jest.fn(),
+      };
+      wrapper.vm.query.path = 'folder';
+      wrapper.vm.query.plugin = 'test';
+      wrapper.vm.localFileInformations = [{ path: 'folder/modelName/file.ext' }];
+
+      await wrapper.vm.openModelFiles();
 
       expect(mock).toHaveBeenCalledTimes(1);
     });
@@ -539,6 +581,7 @@ describe('Test component: FileExplorer', () => {
     it('should update nodes', () => {
       wrapper.vm.fileExplorerRef.getNodeByKey = () => null;
       wrapper.vm.query.path = 'pluginName/modelNamecoucou';
+      wrapper.vm.query.plugin = 'test';
 
       wrapper.vm.initTreeNodes();