+++ changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Add drawer to create diagram from a template.
 * Homepage:
   * Display a warning message when importing a previously imported project.
+* Integrate new diagram management.
 
 ### Changed
 
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Redesign global page.
   * Add buttons to change diagram display (table or grid).
 * Rework and clean all layouts and pages.
+* Have only one configuration file by project to save all diagrams positions.
 
 ### Fixed
 
@@ -36,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Index page:
   * Template grid.
+* Configuration for default path for models.
 
 ## [1.2.0] - 2023/06/07